Squashed commit of the following:
commit 1bb5cb03a1dfa874e255da959437ed293fea8c62
Author: witnessresults <100758276+witnessresults@users.noreply.github.com>
Date:   Sun Sep 25 20:40:04 2022 -0400

    Create .DS_Store

commit 2b8705d76b53c8a7ea21bc629d9a66d1170cb173
Author: Jiang Jiang <gzjjgod@gmail.com>
Date:   Wed Sep 26 11:09:55 2018 -0700

    Update DDK binaries

commit dfcae99e3089ee5c10e39dfdf8ecd08e1635217b
Author: Jiang Jiang <gzjjgod@gmail.com>
Date:   Wed Sep 26 11:07:15 2018 -0700

    Update Dark Mode templete

commit cbb9fa01f2ab5de7a1c37b9dfc41fc719d8f7e47
Author: Jiang Jiang <gzjjgod@gmail.com>
Date:   Wed Sep 26 11:05:38 2018 -0700

    Update Dark Mode templete

commit c22b7b4b830734efd4160e642d40176aac4dcc8e
Author: Jiang Jiang <gzjjgod@gmail.com>
Date:   Wed Sep 26 11:00:05 2018 -0700

    Set deployment target

commit 17f8faf920cc665296a27f3b35b9ab65ac211d84
Author: Jiang Jiang <gzjjgod@gmail.com>
Date:   Wed Sep 26 10:49:39 2018 -0700

    Update build configuration

commit 32a3eebd23a533e98dde403a6e18ff51db71bdce
Author: Jiang Jiang <gzjjgod@gmail.com>
Date:   Tue Oct 27 07:48:12 2015 +0100

    Add download instructions

commit 79cc83dcbfd1169cac1af84cdd18f7f109000966
Author: Jiang Jiang <gzjjgod@gmail.com>
Date:   Tue Aug 11 17:27:51 2015 +0800

    Fix output dictionary file name

commit d89f4192409e2f5f28bc68319d0790ee1a82bdbe
Author: Jiang Jiang <gzjjgod@gmail.com>
Date:   Tue Aug 11 11:45:05 2015 +0800

    Create README.md

commit 6525b8b52e0600a8c31b9ce1895484ec7f67eae1
Author: Jiang Jiang <gzjjgod@gmail.com>
Date:   Tue Aug 11 11:22:09 2015 +0800

    Update with latest DDK and support OS X 10.11 format

commit 6996d9d81d7f5204187d68f9b75c7b42efa5d550
Author: Jiang Jiang <gzjjgod@gmail.com>
Date:   Tue Aug 11 11:10:51 2015 +0800

    Clean up code and make it build on OS X 10.11

commit 9b29f6b278b833b09cc16abb810e9dc55c772fba
Merge: 6c16da1 0dad902
Author: Jjgod Jiang <gzjjgod@gmail.com>
Date:   Sat Oct 1 05:48:20 2011 -0700

    Merge pull request #1 from stewwu/master

    Modified the style in Dictionary Service for better readability in Lion.

commit 0dad902c8ac9a2d506df521764647a5efb834221
Author: Che-Ching Wu <stuart@thepolydice.com>
Date:   Sat Oct 1 11:53:06 2011 +0800

    Modified the style in Dictionary Service for better readability.

commit 6c16da110432edb56e637515494d50ba1d8b4fce
Author: Jjgod Jiang <gzjjgod@gmail.com>
Date:   Wed Jul 27 01:02:24 2011 +0200

    Other changes

commit 9d6fe3c7a0ffabc27ab0e4853e7af78a771a42fc
Author: Jjgod Jiang <gzjjgod@gmail.com>
Date:   Mon Jul 25 21:44:17 2011 +0200

    Use GCD to handle progress report

commit 55634bc1d1a85a1435ae258bdb715869296848a0
Author: Jjgod Jiang <gzjjgod@gmail.com>
Date:   Mon Jul 25 21:16:18 2011 +0200

    Lion support, fix -v 10.6 bug and read progress correctly

+++ sdconv/dict.cpp
@@ -2,8 +2,9 @@
 
 #include "dict.h"
 
-static void parse_description(const char *p, long len, 
-                              std::string &description)
+#include <stdlib.h>
+
+static void parse_description(const char *p, long len, std::string &description)
 {
 	description.clear();
 	const char *p1 = p;
@@ -122,11 +123,11 @@ bool mdk_dict::load_ifo(const gchar *file)
 {
     gchar *buffer;
 
-	if (! g_file_get_contents(file, &buffer, NULL, NULL))
+	if (!g_file_get_contents(file, &buffer, NULL, NULL))
 		return false;
 
 #define DICT_MAGIC_DATA "StarDict's dict ifo file\nversion="
-	if (! g_str_has_prefix(buffer, DICT_MAGIC_DATA))
+	if (!g_str_has_prefix(buffer, DICT_MAGIC_DATA))
     {
 		g_free(buffer);
 		return false;
@@ -157,19 +158,17 @@ bool mdk_dict::load_ifo(const gchar *file)
 	}
 
 	p3 = strchr(p2 + sizeof("\nwordcount=")-1,'\n');
-	gchar *tmpstr = (gchar *)g_memdup(p2 + sizeof("\nwordcount=") - 1, 
-                                      p3 - (p2+sizeof("\nwordcount=") - 1) + 1);
+	gchar *tmpstr = (gchar *)g_memdup(p2 + sizeof("\nwordcount=") - 1, (guint)(p3 - (p2 + sizeof("\nwordcount=") - 1) + 1));
 	tmpstr[p3 - (p2 + sizeof("\nwordcount=") - 1)] = '\0';
-	wordcount = atol(tmpstr);
+	wordcount = (guint32)atol(tmpstr);
 	g_free(tmpstr);
 
 	p2 = strstr(p1,"\nsynwordcount=");
 	if (p2) {
 		p3 = strchr(p2 + sizeof("\nsynwordcount=") - 1, '\n');
-		gchar *tmpstr = (gchar *)g_memdup(p2 + sizeof("\nsynwordcount=") - 1, 
-                                          p3 - (p2 + sizeof("\nsynwordcount=") - 1) + 1);
+		gchar *tmpstr = (gchar *)g_memdup(p2 + sizeof("\nsynwordcount=") - 1, (guint)(p3 - (p2 + sizeof("\nsynwordcount=") - 1) + 1));
 		tmpstr[p3-(p2+sizeof("\nsynwordcount=")-1)] = '\0';
-		synwordcount = atol(tmpstr);
+		synwordcount = (guint32)atol(tmpstr);
 		g_free(tmpstr);
 	} else {
 		synwordcount = 0;
@@ -182,10 +181,9 @@ bool mdk_dict::load_ifo(const gchar *file)
     }
 
     p3 = strchr(p2+ sizeof("\nidxfilesize=")-1,'\n');
-    tmpstr = (gchar *)g_memdup(p2 + sizeof("\nidxfilesize=")-1, 
-                               p3 - (p2 + sizeof("\nidxfilesize=")-1)+1);
-    tmpstr[p3-(p2+sizeof("\nidxfilesize=")-1)] = '\0';
-    index_file_size = atol(tmpstr);
+    tmpstr = (gchar *)g_memdup(p2 + sizeof("\nidxfilesize=") - 1, (guint)(p3 - (p2 + sizeof("\nidxfilesize=") - 1) + 1));
+    tmpstr[p3 - (p2 + sizeof("\nidxfilesize=") - 1)] = '\0';
+    index_file_size = (guint32)atol(tmpstr);
     g_free(tmpstr);
 
     p2 = strstr(p1,"\ndicttype=");
@@ -279,8 +277,8 @@ gchar *mdk_dict::get_entry_data(mdk_entry *entry)
             return NULL;
         }
 
-        guint32 data_size;
-        gint sametypesequence_len = sametypesequence.length();
+        size_t data_size;
+        size_t sametypesequence_len = sametypesequence.length();
         // there have sametypesequence_len char being omitted.
         // Here is a bug fix of 2.4.8, which don't add sizeof(guint32) anymore.
         data_size = idxitem_size + sametypesequence_len; 
@@ -317,7 +315,7 @@ gchar *mdk_dict::get_entry_data(mdk_entry *entry)
         gchar *p1,*p2;
         p1 = data + sizeof(guint32);
         p2 = origin_data;
-        guint32 sec_size;
+        size_t sec_size;
         // copy the head items.
         for (i = 0; i < sametypesequence_len - 1; i++)
         {
